+++ src/linkgraph/linkgraph_gui.cpp
@@ -61,7 +61,7 @@ const uint8 LinkGraphOverlay::LINK_COLOURS[][12] = {
 void LinkGraphOverlay::GetWidgetDpi(DrawPixelInfo *dpi, uint margin) const
 {
 	const NWidgetBase *wi = this->window->GetWidget<NWidgetBase>(this->widget_id);
-	dpi->left = dpi->top = -margin;
+	dpi->left = dpi->top = -static_cast<int>(margin);
 	dpi->width = wi->current_x + 2 * margin;
 	dpi->height = wi->current_y + 2 * margin;
 }
