+++ src/tracerestrict.cpp
@@ -373,11 +373,11 @@ void TraceRestrictProgram::Execute(const Train* v, const TraceRestrictProgramInp
 						const TraceRestrictSlot *slot = TraceRestrictSlot::GetIfValid(GetTraceRestrictValue(item));
 						switch (static_cast<TraceRestrictSlotOccupancyCondAuxField>(GetTraceRestrictAuxField(item))) {
 							case TRSOCAF_OCCUPANTS:
-								result = TestCondition(slot != nullptr ? slot->occupants.size() : 0, condop, value);
+								result = TestCondition(slot != nullptr ? static_cast<int>(slot->occupants.size()) : 0, condop, value);
 								break;
 
 							case TRSOCAF_REMAINING:
-								result = TestCondition(slot != nullptr ? slot->max_occupancy - slot->occupants.size() : 0, condop, value);
+								result = TestCondition(slot != nullptr ? static_cast<int>(slot->max_occupancy - slot->occupants.size()) : 0, condop, value);
 								break;
 
 							default:
@@ -1305,7 +1305,7 @@ CommandCost CmdProgramSignalTraceRestrict(TileIndex tile, DoCommandFlag flags, u
 		return_cmd_error(STR_TRACE_RESTRICT_ERROR_NO_PROGRAM);
 	}
 
-	uint32 offset_limit_exclusive = ((type == TRDCT_INSERT_ITEM) ? 1 : 0);
+	size_t offset_limit_exclusive = ((type == TRDCT_INSERT_ITEM) ? 1 : 0);
 	if (prog) offset_limit_exclusive += prog->items.size();
 
 	if (offset >= offset_limit_exclusive) {
